types

LineBroadenings.Gaussian — Type

Gaussian(μ, σ)

Gaussian distribution, also called normal distribution.

Definition

\[f\left(x ; \mu, \sigma\right)=\frac{1}{\sigma \sqrt{2 \pi}} \exp{\left[-\frac{1}{2} \left(\frac{x-\mu}{\sigma}\right)^{2}\right]}\]

Fields

μ::Real: mean or expectation

σ::Real: standard deviation

External links

Wikipedia

LineBroadenings.Lorentz — Type

Lorentz(μ, γ)

Lorentz distribution, also called Cauchy distribution.

\[f\left(x ; \mu, \gamma\right)=\frac{1}{\pi \gamma\left[1+\left(\frac{x-\mu} {\gamma}\right)^{2}\right]}=\frac{1}{\pi \gamma}\left[\frac{\gamma^{2}} {\left(x-\mu\right)^{2}+\gamma^{2}}\right]\]

Fields

μ::Real: location parameter

γ::Real: scale parameter

External links

Wikipedia

LineBroadenings.Voigt — Type

Voigt(μ, σ, γ)

The Voigt profile is a probability distribution given by a convolution of a Cauchy-Lorentz distribution and a Gaussian distribution.

Definition

\[f(x ;\mu, \sigma, \gamma) \equiv G(x;\mu, \sigma) \otimes L(x;\mu, \gamma)\]

It can be evaluated as

\[f(x ;\mu, \sigma, \gamma)=\frac{\operatorname{Re}[w(z)]}{\sigma \sqrt{2 \pi}}\]

where $\mathrm{Re}[w(z)]$ is the real part of the Faddeeva function evaluated for

\[z=\frac{x-\mu+i \gamma}{\sigma \sqrt{2}}\]

Fields

μ::Real: location parameter

σ::Real: σ in Gaussian distribution

γ::Real: γ in Lorentz distribution

External links

Wikipedia

functions

LineBroadenings.fwhm — Function

fwhm(d::Profile)

The full width at half maximum (FWHM) of d.

LineBroadenings.pdf — Function

pdf(d::Profile, x::Real)

probability density function.

LineBroadenings.peak — Function

peak(d::Profile)

maximum of the pdf of d.